So many things - new grid layout based on 6 columns Added screenie for screen shots, with range of potatoe-shop style colour control Made tabgroup heading look nicer Added more settings and added missing bits for SettingOptionFloatRange Fixed web server running out of async tcp bits - now streaming back update results Adde ServerEvents for communicating with the webserver and live web pages

+++ platformio/src/ui/controls/ui_control_slider.cpp
@@ -185,13 +185,13 @@ bool ui_control_slider::process_touch(touch_event_t touch_event)
 			// 2. snap it to nearest step
 			//    e.g. steps = 10, halfStep = 5
 			int steps = int((raw - value_min + value_step / 2.0f) / value_step);
-			int snapped = value_min + steps * value_step;
+			float snapped = value_min + steps * value_step;
 
 			// 3. clamp just in case
 			current_value = constrain(snapped, value_min, value_max);
 
 			if (value_type == VALUE_TYPE::INT)
-				static_cast<SettingsOptionInt *>(setting_option)->update(current_value);
+				static_cast<SettingsOptionInt *>(setting_option)->update((int)current_value);
 			else if (value_type == VALUE_TYPE::FLOAT)
 				static_cast<SettingsOptionFloat *>(setting_option)->update(current_value);
